organizing codebase
