+++ Podcast/ActionSheetViewController.swift
@@ -1,291 +1,8 @@
 
 import UIKit
 
-protocol ActionSheetTableViewCellProtocol: class {
-    static var identifier: String { get }
-    static var cellHeight: CGFloat { get }
-    func setup(withOption option: ActionSheetOptionType)
-}
-
-class ActionSheetStandardTableViewCell: UITableViewCell {
-    
-    let padding: CGFloat = 18
-    var leftPadding: CGFloat = 50
-    var titleLabel: UILabel!
-    var iconImage: UIImageView!
-    
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        backgroundColor = .offWhite
-        separatorInset = UIEdgeInsets(top: 0, left: padding, bottom: 0, right: 0)
-        
-        titleLabel = UILabel()
-        titleLabel.numberOfLines = 3
-        titleLabel.font = ._14RegularFont()
-        addSubview(titleLabel)
-        
-        iconImage = UIImageView()
-        addSubview(iconImage)
-        
-        iconImage.snp.makeConstraints { make in
-            make.leading.equalToSuperview().inset(padding)
-            make.centerY.equalToSuperview()
-        }
-        
-        titleLabel.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.leading.equalTo(leftPadding)
-            make.trailing.equalToSuperview().inset(padding)
-        }
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setup(withOption option: ActionSheetOptionType) {
-        titleLabel.text = option.title
-        titleLabel.textColor = option.titleColor
-        iconImage.snp.remakeConstraints { make in
-            make.leading.equalToSuperview().inset(padding)
-            make.centerY.equalToSuperview()
-            make.size.equalTo(option.iconImage?.size ?? 0)
-        }
-        iconImage.image = option.iconImage
-    }
-    
-}
-
-class ActionSheetMoreOptionsStandardTableViewCell: ActionSheetStandardTableViewCell, ActionSheetTableViewCellProtocol {
-
-    static var identifier: String = "actionSheetStandardTableViewCellIdentifier"
-    static var cellHeight: CGFloat = 58
-
-
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-class ActionSheetRecastDescriptionTableViewCell: ActionSheetStandardTableViewCell, ActionSheetTableViewCellProtocol {
-    static var identifier: String = "actionSheetRecastDescriptionTableViewCellIdentifier"
-    static var cellHeight: CGFloat = 70
-
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-protocol ActionSheetPlayerControlsTableViewCellDelegate: class {
-    func didPressSegmentedControlForTrimSilence(selected: Bool)
-}
-
-class ActionSheetPlayerControlsTableViewCell: UITableViewCell, ActionSheetTableViewCellProtocol {
-
-    let padding: CGFloat = 18
-    let leftPadding: CGFloat = 50
-
-    static var identifier: String = "actionSheetPlayerControlsTableViewCellIdentifier"
-    static var cellHeight: CGFloat = 58
-
-    var switchControl: UISwitch!
-    var titleLabel: UILabel!
-    weak var delegate: ActionSheetPlayerControlsTableViewCellDelegate?
-
-    override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-
-        backgroundColor = .offWhite
-        separatorInset = UIEdgeInsets(top: 0, left: padding, bottom: 0, right: 0)
-
-        titleLabel = UILabel()
-        titleLabel.numberOfLines = 3
-        titleLabel.font = ._14RegularFont()
-        addSubview(titleLabel)
-
-        switchControl = UISwitch()
-        switchControl.onTintColor = .sea
-        switchControl.addTarget(self, action: #selector(segmentedControlPress), for: .valueChanged)
-        addSubview(switchControl)
-
-        switchControl.snp.makeConstraints { make in
-            make.leading.equalToSuperview().inset(padding)
-            make.centerY.equalToSuperview()
-        }
-
-        titleLabel.snp.makeConstraints { make in
-            make.centerY.equalToSuperview()
-            make.leading.equalTo(switchControl.snp.trailing).offset(padding)
-            make.trailing.equalToSuperview().inset(padding)
-        }
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    func setup(withOption option: ActionSheetOptionType) {
-        titleLabel.text = option.title
-        titleLabel.textColor = option.titleColor
-        switch(option) {
-        case .playerSettingsTrimSilence(let selected):
-            switchControl.isOn = selected
-        default:
-            switchControl.isOn = false
-        }
-    }
-
-    @objc func segmentedControlPress() {
-        delegate?.didPressSegmentedControlForTrimSilence(selected: switchControl.isOn)
-    }
-}
-
-
-
-class ActionSheetHeaderView: UIView {
-    
-    var imageView: ImageView!
-    var titleLabel: UILabel!
-    var descriptionLabel: UILabel!
-    
-    let episodeNameLabelY: CGFloat = 27
-    let episodeNameLabelX: CGFloat = 86.5
-    let episodeNameLabelRightX: CGFloat = 21
-    let episodeNameLabelHeight: CGFloat = 18
-    let descriptionLabelX: CGFloat = 86.5
-    let descriptionLabelY: CGFloat = 47.5
-    var descriptionLabelHeight: CGFloat = 14.5
-    var padding: CGFloat = 18
-    let smallPadding: CGFloat = 2
-    var imageViewSize: CGFloat = 60
-    
-    init(frame: CGRect, image: UIImage, title: String, description: String) {
-        super.init(frame: frame)
-        
-        backgroundColor = .offWhite
-        
-        imageView = ImageView()
-        titleLabel = UILabel()
-        descriptionLabel = UILabel()
-
-        titleLabel.font = ._14SemiboldFont()
-        titleLabel.textColor = .offBlack
-        titleLabel.numberOfLines = 2
-
-        descriptionLabel.font = ._12RegularFont()
-        descriptionLabel.textColor = .charcoalGrey
-        descriptionLabel.numberOfLines = 2
-
-        imageView.image = image
-        titleLabel.text = title
-        descriptionLabel.text = description
-        
-        addSubview(imageView)
-        addSubview(titleLabel)
-        addSubview(descriptionLabel)
-        
-        imageView.snp.makeConstraints { make in
-            make.top.equalToSuperview().inset(padding)
-            make.leading.equalToSuperview().inset(padding)
-            make.size.equalTo(imageViewSize)
-        }
-        
-        titleLabel.snp.makeConstraints { make in
-            make.leading.equalTo(imageView.snp.trailing).offset(padding)
-            make.trailing.equalToSuperview().inset(padding)
-            make.top.equalTo(imageView.snp.top)
-        }
-        
-        descriptionLabel.snp.makeConstraints { make in
-            make.top.equalTo(titleLabel.snp.bottom).offset(smallPadding)
-            make.leading.equalTo(titleLabel.snp.leading)
-            make.trailing.equalTo(titleLabel.snp.trailing)
-        }
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-enum ActionSheetOptionType {
-    case download(selected: Bool)
-    case bookmark(selected: Bool)
-    case recommend(selected: Bool)
-    case listeningHistory
-    case shareEpisode
-    case confirmShare
-    case deleteShare
-    case recastDescription
-    case playerSettingsTrimSilence(selected: Bool)
-    
-    var title: String {
-        switch (self) {
-        case .download(let selected):
-            return selected ? "Remove download" : "Download this episode"
-        case .bookmark(let selected):
-            return selected ? "Remove save" : "Save for later"
-        case .recommend(let selected):
-            return selected ? "Undo recast" : "Recast this episode"
-        case .listeningHistory:
-            return "Remove from Listening History"
-        case .shareEpisode:
-            return "Share this episode"
-        case .confirmShare:
-            return "Click to confirm share"
-        case .deleteShare:
-            return "Delete shared episode"
-        case .recastDescription:
-            return "When you recast a podcast episode, that episode is shared with your followers on their feed."
-        case .playerSettingsTrimSilence:
-            return "Trim silent parts of episode"
-        }
-    }
-    
-    var iconImage: UIImage? {
-        switch(self) {
-        case .download(let selected):
-            return selected ? #imageLiteral(resourceName: "download_remove") : #imageLiteral(resourceName: "download")
-        case .bookmark(let selected):
-            return selected ? #imageLiteral(resourceName: "bookmark_feed_icon_selected") : #imageLiteral(resourceName: "bookmark_feed_icon_unselected")
-        case .recommend(let selected):
-            return selected ? #imageLiteral(resourceName: "repost_selected") : #imageLiteral(resourceName: "repost")
-        case .listeningHistory, .deleteShare:
-            return #imageLiteral(resourceName: "failure_icon")
-        case .shareEpisode, .confirmShare:
-            return #imageLiteral(resourceName: "iShare")
-        case .recastDescription:
-            return #imageLiteral(resourceName: "repost_selected")
-        default:
-            return nil
-        }
-    }
-
-    var titleColor: UIColor {
-        return .charcoalGrey
-    }
-
-    var cell: ActionSheetTableViewCellProtocol.Type {
-        switch(self) {
-        case .recastDescription:
-            return ActionSheetRecastDescriptionTableViewCell.self
-        case .playerSettingsTrimSilence:
-            return ActionSheetPlayerControlsTableViewCell.self
-        default:
-            return ActionSheetMoreOptionsStandardTableViewCell.self
-        }
-    }
-}
-
+// Make an action class for each action sheet option
+// upon tap -> complete action
 class ActionSheetOption {
     
     var type: ActionSheetOptionType
@@ -298,6 +15,7 @@ class ActionSheetOption {
     
 }
 
+// Make a header class if you want a header for your action sheet
 class ActionSheetHeader {
     
     var image: UIImage
